requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
streamlit
beautifulsoup4
langchain-ollama